WallHitSpring Patch
With this modification, it will prevent the camera from clipping trough walls and floors. When using the shooter controller, the camera always goes trough the floor when trying to look up.

Lots of people were complaining in the discord about this so I decided to fix it because I had this problem too.